refactor: remove old UserAvatar and Avatar components
