--- src/components/shared/Avatar/index.ts
@@ -1,1 +1,0 @@
-export { default, type Props as AvatarProps } from './Avatar.tsx';
