+++ src/stories/common/MemberSignature.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import MemberSignature from '~v5/common/MemberSignature/index.ts';
-import { Avatar2 } from '~v5/shared/Avatar/Avatar.tsx';
+import { Avatar } from '~v5/shared/Avatar/Avatar.tsx';
 
 import type { Meta, StoryObj } from '@storybook/react';
 
@@ -34,7 +34,7 @@ export const Base: StoryObj<typeof MemberSignature> = {
   args: {
     hasSigned: true,
     children: (
-      <Avatar2
+      <Avatar
         size={20}
         address="0xb77D57F4959eAfA0339424b83FcFaf9c15407461"
         alt="user avatar"
